+++ src/types.h
@@ -26,6 +26,8 @@ typedef union binary64_alias {
 	double f;
 } binary64_alias;
 
+typedef uint16_t cbor_fp16i;
+
 void php_cbor_minit_types();
 
 /* undefined */
@@ -39,10 +41,18 @@ zend_string *cbor_get_xstring_value(zval *value);
 /* floatx */
 zend_object *cbor_floatx_create(zend_class_entry *ce);
 bool cbor_floatx_set_value(zend_object *obj, zval *value, uint32_t raw);
-size_t cbor_floatx_get_value(zend_object *obj, char *out);
-double cbor_from_float16(uint16_t value);
-float cbor_to_float32(double value);
-uint16_t cbor_to_float16(double value);
+size_t cbor_floatx_get_value_be(zend_object *obj, char *out);
+float cbor_float32_get_value(zend_object *obj);
+double cbor_from_fp16i(cbor_fp16i value);
+double cbor_from_fp32(float value);
+float cbor_to_fp32(double value);
+int test_fp32_size(float value);
+int test_fp64_size(double value);
+void php_cbor_minit_types_float_cast();
+cbor_fp16i cbor_float_32_to_16(float value);
+cbor_fp16i cbor_float_32_to_16_cast(float value);
+cbor_fp16i cbor_float_64_to_16(double value);
+cbor_fp16i cbor_float_64_to_16_cast(double value);
 
 /* decoder */
 void php_cbor_minit_decoder();