+++ stubs/.phpstorm.meta.php
@@ -9,7 +9,7 @@ namespace PHPSTORM_META {
 /* encode flags start */
 CBOR_SELF_DESCRIBE | CBOR_BYTE | CBOR_TEXT | CBOR_INT_KEY | CBOR_KEY_BYTE
  | CBOR_KEY_TEXT | CBOR_UNSAFE_TEXT | CBOR_MAP_NO_DUP_KEY | CBOR_FLOAT16
- | CBOR_FLOAT32
+ | CBOR_FLOAT32 | CBOR_CDE
 /* encode flags end */
     );
     registerArgumentsSet('CborDecodeFlags',